Cleaned compilation warnings from VC++.
git-svn-id: http://svn.osgeo.org/geos/trunk@1997 5242fede-7e19-0410-aef8-94bd7d2200fb

+++ source/headers/geos/io/ByteOrderDataInStream.inl
@@ -76,7 +76,7 @@ ByteOrderDataInStream::readLong()
 	stream->read(reinterpret_cast<char *>(buf), 8);
 	if ( stream->eof() )
 		throw  ParseException("Unexpected EOF parsing WKB");
-	return ByteOrderValues::getLong(buf, byteOrder);
+	return static_cast<long>(ByteOrderValues::getLong(buf, byteOrder));
 }
 
 INLINE double
